Changes made few states and a function moved to context
