+++ src/CryptoContext.js
@@ -1,5 +1,7 @@
+import axios from 'axios';
 import React, { useContext, useEffect, useState } from 'react';
 import { createContext } from 'react';
+import { CoinList } from './config/API';
 
 const Crypto = createContext();
 
@@ -7,6 +9,17 @@ const CryptoContext = ({children}) => {
 
   const [currency, setCurrency] = useState("INR");
   const [symbol, setSymbol] = useState("₹");
+  const [coins, setCoins] = useState([]);
+  const [loading, setLoading] = useState(false);
+
+  const fetchCoins = async () => {
+    setLoading(true);
+    const {data} = await axios.get(CoinList(currency));
+    // this is a bad technique follow the courses technique
+    setCoins(data);
+    setLoading(false);
+}
+
 
   useEffect(()=>{
     if(currency === "INR") setSymbol("₹")
@@ -14,7 +27,7 @@ const CryptoContext = ({children}) => {
   },[currency])
 
   return (
-    <Crypto.Provider value={{currency, symbol, setCurrency}}>
+    <Crypto.Provider value={{currency, symbol, setCurrency, coins, loading, fetchCoins}}>
       {children}
     </Crypto.Provider>
   )
